requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1
requests==2.32.3
Pillow==10.4.0
numpy==2.1.3
psutil==6.1.0
pynvml==11.5.3
pypdf==5.1.0
PyMuPDF==1.24.13
transformers==4.46.3
tokenizers==0.20.3
torch
oss2==2.19.2
prometheus-client==0.21.0
fastmcp
